add of some unit test
